Handle empty response in fetchProducts()
Added a check to handle the case when the API response is successful but the data array is empty. If the response contains no products, an error is thrown to indicate that no products were found. This improves the robustness of the fetchProducts() method by providing better feedback to the caller when no products are returned from the API.